Fix memory leaks in biteopt. I never figured out precisely what was wrong, but switching to std::vector instead of manual new/delete did the trick and is preferable to me anyway. Fixed a bunch of g++ and ASAN warnings while I'm on linux too.
git-svn-id: file:///home/runner/work/tom7misc-sync-action/tom7misc-sync-action/upstream/trunk@6133 c7e11af0-4c7a-11f1-9edd-117cb6f09d55

+++ cc-lib/opt/large-optimizer.h
@@ -139,7 +139,7 @@ struct LargeOptimizer {
   struct HashArg {
     size_t operator ()(const arg_type &arg) const {
       uint64_t result = 0xCAFEBABE;
-      for (int i = 0; i < arg.size(); i++) {
+      for (int i = 0; i < (int)arg.size(); i++) {
         result *= 0x113399557799;
         // This seems to be the most standards-compliant way to get
         // the bytes of a double?
@@ -268,7 +268,7 @@ inline void LargeOptimizer<CACHE>::Run(
 
   std::vector<int> optcounts(arginfos.size(), 0);
   std::vector<int> indices(arginfos.size());
-  for (int i = 0; i < indices.size(); i++) indices[i] = i;
+  for (int i = 0; i < (int)indices.size(); i++) indices[i] = i;
 
   auto Shuffle = [&RandTo32](std::vector<int> *v) {
       if (v->size() <= 1) return;
@@ -285,7 +285,7 @@ inline void LargeOptimizer<CACHE>::Run(
       Shuffle(&indices);
       std::vector<int> subset;
       subset.reserve(size);
-      for (int i = 0; i < indices.size() && i < size; i++)
+      for (int i = 0; i < (int)indices.size() && i < size; i++)
         subset.push_back(indices[i]);
 
       // Keep track of how many times we've used them.
